Status DOT Bug FIx
DOT value was put in, but did not give any DOT to the enemy because of int float shenanigans. When the duration of a status effect wears off, removing the status effect itself from the list while doing a foreach loop on that list creates an error where the collection was modified and the enumeration operation cannot continue properly. Created another list to store which status effects to remove after the foreach loop.